refactor(insights): hoist Funnel to shared tabs/components
Move the three byte-identical tab-local Funnel components (gates, prompts,
conversion) into a single canonical Funnel in tabs/components/ with a generic
FunnelStage type and newspack-insights__funnel* BEM classes. Funnel styles
move to _insights-charts.scss; per-tab funnel blocks removed from gates.scss,
prompts.scss, conversion.scss.
