+++ plugins/newspack-plugin/src/wizards/insights/tabs/components/Funnel.test.tsx
