+++ plugins/newspack-plugin/src/wizards/insights/tabs/prompts/HowReadersConvertSection.tsx
@@ -18,7 +18,7 @@ import { __ } from '@wordpress/i18n';
  */
 import type { PromptsWindow } from '../../api/prompts';
 import SectionHeading from '../components/SectionHeading';
-import Funnel from './viz/Funnel';
+import Funnel from '../components/Funnel';
 import DistributionTable from './viz/DistributionTable';
 import SectionState from './SectionState';
